+++ cardinal-components-base/src/main/java/dev/onyxstudios/cca/api/v3/component/AutoSyncedComponent.java
@@ -24,24 +24,19 @@ package dev.onyxstudios.cca.api.v3.component;
 
 import com.demonwav.mcdev.annotations.CheckEnv;
 import com.demonwav.mcdev.annotations.Env;
-import nerdhub.cardinal.components.api.component.Component;
-import net.minecraft.entity.Entity;
-import net.minecraft.nbt.CompoundTag;
+import dev.onyxstudios.cca.api.v3.component.sync.ComponentPacketWriter;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.util.Util;
-import net.minecraft.world.World;
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Contract;
 
 /**
  * A {@link Component} implementing this interface will have its data automatically
  * synchronized with players watching its provider.
  *
- * @see ComponentKey#sync(Object)
+ * @deprecated implement {@link dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent} instead
  */
-// TODO consider renaming packet methods to writeSyncPacket and applySyncPacket
-public interface AutoSyncedComponent extends ComponentV3 {
+@Deprecated
+public interface AutoSyncedComponent extends dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent {
     /**
      * The default sync operation.
      *
@@ -50,13 +45,9 @@ public interface AutoSyncedComponent extends ComponentV3 {
      */
     int FULL_SYNC = 0;
 
-    /**
-     * @deprecated use/override {@link #shouldSyncWith(ServerPlayerEntity, int)} instead
-     */
-    @Deprecated
-    @ApiStatus.ScheduledForRemoval
+    @Override
     default boolean shouldSyncWith(ServerPlayerEntity player) {
-        return true;
+        return this.shouldSyncWith(player, FULL_SYNC);
     }
 
     /**
@@ -64,69 +55,38 @@ public interface AutoSyncedComponent extends ComponentV3 {
      * should be immediately sent to {@code player}.
      *
      * @param player potential recipient of a synchronization packet
-     * @param syncOp the specific sync operation to be performed, as passed in {@link ComponentKey#sync(Object, int)}
      * @return {@code true} if synchronization with the {@code player} should occur,
      * {@code false} otherwise
      */
     @Contract(pure = true)
     default boolean shouldSyncWith(ServerPlayerEntity player, int syncOp) {
-        // calling the deprecated overload for backward compatibility
-        return this.shouldSyncWith(player);
+        return dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent.super.shouldSyncWith(player);
+    }
+
+    @Override
+    default void writeSyncPacket(PacketByteBuf buf, ServerPlayerEntity recipient) {
+        this.writeToPacket(buf, recipient, FULL_SYNC);
     }
 
     /**
-     * @deprecated use/override {@link #writeToPacket(PacketByteBuf, ServerPlayerEntity, int)}
-     */
-    @ApiStatus.ScheduledForRemoval
-    @Deprecated
-    default void writeToPacket(PacketByteBuf buf, ServerPlayerEntity recipient) {
-        buf.writeCompoundTag(Util.make(new CompoundTag(), this::writeToNbt));
-    }
-
-    /**
-     * Writes this component's data to {@code buf}.
-     *
-     * <p>Implementations may choose to change the data written to the packet based on the {@code recipient}
-     * or on the {@code syncOp}, or they may safely ignore both parameters.
-     * A {@code syncOp} value of {@code 0} triggers the base full synchronization behaviour.
-     * Other values have a meaning specific to the component implementation.
-     * Typical uses of this parameter include limiting the amount of data being synced (eg. by identifying a
-     * specific field to write), or triggering different events on the client (similar to {@link World#sendEntityStatus(Entity, byte)})
-     *
-     * @param buf       the buffer to write the data to
-     * @param recipient the player to which the packet will be sent
-     * @param syncOp    the specific sync operation to be performed, as passed in {@link ComponentKey#sync(Object, int)}
-     * @implSpec The default implementation writes the whole NBT representation
-     * of this component to the buffer using {@link #writeToNbt(CompoundTag)}.
-     * @implNote The default implementation should generally be overridden.
-     * The serialization done by the default implementation sends possibly hidden
-     * information to clients, uses a wasteful data format, and does not support
-     * any optimization such as incremental updates. Implementing classes can
-     * nearly always provide a better implementation.
-     * @see ComponentKey#sync(Object)
-     * @see ComponentKey#sync(Object, int)
-     * @see #readFromPacket(PacketByteBuf)
+     * @deprecated override {@link dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent#writeSyncPacket(PacketByteBuf, ServerPlayerEntity)} instead.
+     * <p>If you use {@code syncOp}, pass the extra information through a custom {@link ComponentPacketWriter} in {@link ComponentKey#sync(Object, ComponentPacketWriter)}
      */
     @Contract(mutates = "param1")
     default void writeToPacket(PacketByteBuf buf, ServerPlayerEntity recipient, int syncOp) {
-        // calling the deprecated overload for backward compatibility
-        this.writeToPacket(buf, recipient);
+        dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent.super.writeSyncPacket(buf, recipient);
+    }
+
+    @Override
+    default void applySyncPacket(PacketByteBuf buf) {
+        this.readFromPacket(buf);
     }
 
     /**
-     * Reads this component's data from {@code buf}.
-     *
-     * @implSpec The default implementation converts the buffer's content
-     * to a {@link CompoundTag} and calls {@link #fromTag(CompoundTag)}.
-     * @implNote any implementing class overriding {@link #writeToPacket(PacketByteBuf, ServerPlayerEntity)}
-     * such that it uses a different data format must override this method.
-     * @see #writeToPacket(PacketByteBuf, ServerPlayerEntity)
+     * @deprecated override {@link dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent#applySyncPacket(PacketByteBuf)} instead
      */
     @CheckEnv(Env.CLIENT)
     default void readFromPacket(PacketByteBuf buf) {
-        CompoundTag tag = buf.readCompoundTag();
-        if (tag != null) {
-            this.readFromNbt(tag);
-        }
+        dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent.super.applySyncPacket(buf);
     }
 }