+++ cardinal-components-base/src/main/java/nerdhub/cardinal/components/api/component/extension/package-info.java
@@ -23,6 +23,7 @@
 /**
  * Special component traits
  */
+@Deprecated
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
 package nerdhub.cardinal.components.api.component.extension;