+++ cardinal-components-entity/src/main/java/nerdhub/cardinal/components/CardinalComponentsEntity.java
@@ -22,8 +22,8 @@
  */
 package nerdhub.cardinal.components;
 
-import dev.onyxstudios.cca.api.v3.component.AutoSyncedComponent;
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
+import dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent;
 import dev.onyxstudios.cca.internal.base.ComponentsInternals;
 import dev.onyxstudios.cca.internal.base.InternalComponentProvider;
 import nerdhub.cardinal.components.api.ComponentRegistry;
@@ -106,7 +106,7 @@ public final class CardinalComponentsEntity {
                             componentType.maybeGet(context.getPlayer().world.getEntityById(entityId))
                                 .ifPresent(c -> {
                                     if (c instanceof AutoSyncedComponent) {
-                                        ((AutoSyncedComponent) c).readFromPacket(copy);
+                                        ((AutoSyncedComponent) c).applySyncPacket(copy);
                                     } else if (c instanceof SyncedComponent) {
                                         ((SyncedComponent) c).processPacket(context, copy);
                                     }